refactor(server): extract agent init/builders/reload to server/agent_init.py (ADR 0023 phase 2)
Final phase-2 backend extraction. Moves agent composition out of the root into
server/agent_init.py (27 functions):
- _init_langgraph_agent + the ten _build_* component builders (knowledge,
  skills, MCP, plugins, checkpointer, inbox, activity, telemetry, workflow,
  scheduler)
- the checkpoint-prune + thread-retire background loops
- plugin-host wiring (_populate_plugin_host, _register_plugin_subagents,
  _reload_plugin_surfaces, _plugin_agent_invoke)
- _reload_langgraph_agent (hot reload) + the offload helper _run_on_server_loop
- the operator-console settings callbacks (_apply_settings_changes,
  _build_settings_callbacks, _sync_autostart_with_config)

The builders read/mutate the shared runtime.state.STATE; the four __init__
symbols they need (agent_name, AGENT_NAME_ENV, _event_bus, _bundle_root) are
imported from `server` — all defined before the re-export line that triggers
this import, so it's not a cycle. server/__init__.py re-exports all 27 names
(verified same-object) and drops ~1,135 lines: the composition root is now
~1,354 lines, down from 3,353 before phase 1. The remaining bulk is _main's
route handlers — ADR 0023 phase 3.

No test monkeypatches a builder (they call server._build_X directly, or patch
external deps / STATE), so the inter-builder calls staying within agent_init.py
is behavior-preserving.

Verified: 1000 tests green; live smoke via `python -m server` — boot ran every
builder cleanly (checkpointer, hybrid knowledge store, scheduler, 3 plugins +
subagent registration, skills, workflows, telemetry, graph compiled), a chat
turn, and a config-driven hot reload (_reload_langgraph_agent -> "agent
reloaded", graph recompiled, post-reload turn OK).
